Input and output method types directly defined in the method files, then re-exported
