+++ packages/sdk/src/dataProtector/sharing/getSubscribers.ts
@@ -3,27 +3,34 @@ import {
   throwIfMissing,
 } from '../../utils/validators.js';
 import {
-  GetSubscribersResponse,
-  GraphQLResponseSubscribers,
+  Address,
+  CollectionTokenIdParam,
   SubgraphConsumer,
-  SubscribeParams,
-  Subscriber,
-} from '../types.js';
+} from '../types/shared.js';
 import { getCollectionSubscribers } from './subgraph/getCollectionSubscribers.js';
+
+type Subscriber = {
+  address: Address;
+  endSubscriptionTimestamp: number;
+};
+
+export type GetSubscribersResponse = {
+  subscribers: Subscriber[];
+};
+
 export const getSubscribers = async ({
   graphQLClient = throwIfMissing(),
   collectionTokenId,
-}: SubscribeParams & SubgraphConsumer): Promise<GetSubscribersResponse> => {
+}: SubgraphConsumer &
+  CollectionTokenIdParam): Promise<GetSubscribersResponse> => {
   const vCollectionId = positiveNumberSchema()
     .required()
     .label('collectionId')
     .validateSync(collectionTokenId);
-
-  const getSubscribersQueryResponse: GraphQLResponseSubscribers =
-    await getCollectionSubscribers({
-      collectionTokenId: vCollectionId,
-      graphQLClient,
-    });
+  const getSubscribersQueryResponse = await getCollectionSubscribers({
+    collectionTokenId: vCollectionId,
+    graphQLClient,
+  });
 
   const subscribers: Subscriber[] =
     getSubscribersQueryResponse.collectionSubscriptions.map((item) => ({
