+++ packages/sdk/src/dataProtector/sharing/setSubscriptionParams.ts
@@ -1,16 +1,19 @@
 import { WorkflowError } from '../../utils/errors.js';
 import { throwIfMissing } from '../../utils/validators.js';
-import {
-  SetSubscriptionParams,
-  SetSubscriptionParamsResponse,
-} from '../types.js';
+import { SuccessWithTransactionHash } from '../types/shared.js';
 import { getSharingContract } from './smartContract/getSharingContract.js';
+
+export type SetSubscriptionParams = {
+  collectionTokenId: number;
+  priceInNRLC: bigint;
+  durationInSeconds: number;
+};
 
 export const setSubscriptionParams = async ({
   collectionTokenId = throwIfMissing(),
   priceInNRLC = throwIfMissing(),
   durationInSeconds = throwIfMissing(),
-}: SetSubscriptionParams): Promise<SetSubscriptionParamsResponse> => {
+}: SetSubscriptionParams): Promise<SuccessWithTransactionHash> => {
   try {
     //TODO:Input validation
     const sharingContract = await getSharingContract();
