+++ packages/sdk/src/dataProtector/sharing/smartContract/approveCollectionContract.ts
@@ -1,6 +1,6 @@
 import { ethers } from 'ethers';
 import { getPocoRegistryContract } from '../../smartContract/getPocoRegistryContract.js';
-import type { Address, AddressOrENS } from '../../types.js';
+import type { Address, AddressOrENS } from '../../types/shared.js';
 
 export async function approveCollectionContract({
   protectedDataAddress,
