+++ packages/sdk/src/index.ts
@@ -1,3 +1,3 @@
-export * from './dataProtector/types.js';
+export * from './dataProtector/types/index.js';
 export { IExecDataProtector } from './dataProtector/IExecDataProtector.js';
 export { getWeb3Provider } from './utils/getWeb3Provider.js';
