+++ packages/sdk/src/utils/getWeb3Provider.ts
@@ -1,5 +1,5 @@
 import { getSignerFromPrivateKey } from 'iexec/utils';
-import { Web3SignerProvider } from '../dataProtector/types.js';
+import { Web3SignerProvider } from '../dataProtector/types/index.js';
 
 export const getWeb3Provider = (privateKey: string): Web3SignerProvider =>
   getSignerFromPrivateKey('bellecour', privateKey);
